feature: 상품 상세 조회 API 추가 및 연결
