+++ README.md
@@ -72,16 +72,7 @@ Package Manager: `Yarn`
 
 ## 프로젝트 실행 방법
 
-1. Env 세팅
-   root directory
-
-```shell
-NEXT_PUBLIC_API =
-```
-
-<br>
-
-2. 프로젝트 실행
+1. 프로젝트 실행
 
 ```shell
 # 필요한 모듈 설치
