+++ next.config.mjs
@@ -3,7 +3,7 @@ const nextConfig = {
   reactStrictMode: true,
   distDir: 'build', // 정적 웹 사이트 구축을 위한 설정
   // output: 'export', // 정적 웹 사이트 구축을 위한 설정
-  trailingSlash: true, // 정적 웹 사이트 구축을 위한 설정
+  // trailingSlash: true, // 정적 웹 사이트 구축을 위한 설정
   images: {
     unoptimized: true, // 정적 웹 사이트 구축을 위한 설정
     remotePatterns: [
